:recycle: :art: unify main layout and styles, optimize page formatting
- Consolidated main page structure and CSS
- Enhanced formatting and consistency across pages
