+++ pages/login.vue
@@ -3,11 +3,7 @@
 </script>
 
 <template>
-  <v-app>
-    <v-container>
 
-    </v-container>
-  </v-app>
 </template>
 
 <style scoped>
